Split piece code to piece and moves
