+++ src/piece.cpp
@@ -1,23 +1,32 @@
 #include "piece.h"
 
-//Convert notation through indexing array
-const std::string index_to_string[64] = {
-    "a1", "b1", "c1", "d1", "e1", "f1", "g1", "h1",
-    "a2", "b2", "c2", "d2", "e2", "f2", "g2", "h2",
-    "a3", "b3", "c3", "d3", "e3", "f3", "g3", "h3",
-    "a4", "b4", "c4", "d4", "e4", "f4", "g4", "h4",
-    "a5", "b5", "c5", "d5", "e5", "f5", "g5", "h5",
-    "a6", "b6", "c6", "d6", "e6", "f6", "g6", "h6",
-    "a7", "b7", "c7", "d7", "e7", "f7", "g7", "h7",
-    "a8", "b8", "c8", "d8", "e8", "f8", "g8", "h8"
-};
+std::ostream &operator<<(std::ostream &out, PieceType piece) {
+    switch (piece) {
+        case NoPieceType:
+            return out << "No Piece";
+        case Pawn:
+            return out << "Pawn";
+        case Knight:
+            return out << "Knight";
+        case Bishop:
+            return out << "Bishop";
+        case Rook:
+            return out << "Rook";
+        case Queen:
+            return out << "Queen";
+        case King:
+            return out << "King";
+        default:
+            return out;
+    }
+}
 
-const std::string flag_to_string[3] = {
-    "NULL", "", "(capture)"
-};
-
-std::ostream &operator<<(std::ostream &out, Move move) {
-    std::string str;
-    str += move.UCI() + ' ' + flag_to_string[move.flag() >> 12];
-    return out << str;
+std::ostream &operator<<(std::ostream &out, Piece piece) {
+    switch (getPieceColor(piece)) {
+        case WHITE:
+            out << "White ";
+        case BLACK:
+            out << "Black ";
+    }
+    return out << getPieceType(piece);
 }
